This change enforces static type checking using the QuantMatrix type to avoid cyclic imports, but causes pytest to fail.
Changing QuantType to Any when type checking isn't happening allows for the tests to pass, and the static type checking to pass without resorting to using Any.

This is not really elegant, but it seems to work. Im not sure if there is some flag in pytest that will fix this