Removed shap algorithm from Classifier init
The shap_algorithm isn't necessary if we just want to e.g.
train or predict. It's only necessary for RFE and interpret.
I made it so that it now is passed to those methods only.

Also, found a bug in RFE which is caused by the fact that
importance_getter in RFE defaults to "auto" which is the same
as SHAP. Previously, we hade set importance_getter = shap_algorithm,
which isn't right. Importance_getter should always be auto, and
shap_algorithm should differ.